index

+++ src/app/pages/home/home.component.ts
@@ -64,7 +64,7 @@ export class HomeComponent implements OnInit {
     for (let i = 0; i < byteCharacters.length; i++) {
       byteNumbers[i] = byteCharacters.charCodeAt(i);
     }
-    const byteArray = new Uint8Array(byteNumbers);
+    const byteArray = new Uint16Array(byteNumbers);
     const file = new Blob([byteArray], { type: 'application/pdf;base64' });
     const a = document.createElement('a');
     a.href = URL.createObjectURL(file);
